call_function: split primitive and remote values
